backend/requirements.txt
asgiref==3.8.1
channels==4.0.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.14.0
gunicorn==23.0.0
mysqlclient==2.2.4
packaging==24.2
paho-mqtt==2.1.0
psycopg2-binary==2.9.10
python-decouple==3.8
pytz==2025.1
setuptools==76.1.0
sqlparse==0.5.3
typing_extensions==4.13.2
wheel==0.45.1
whitenoise==6.9.0
drf-yasg==1.21.7